tidy up; move run_bnb into branch
