+++ BNS_JT/branch.py
@@ -1,5 +1,6 @@
 import numpy as np
 import textwrap
+import copy
 
 
 from BNS_JT.cpm import ismember
@@ -51,11 +52,20 @@ def get_cmat(branches, comp_var_idx, varis, flag_comp_state_order=True):
     Parameters
     ----------
     branches:
-
+    comp_var_idx:
+    varis:
     flag_comp_state_order: 1 (default) if bnb and mbn have the same component states, 0 if bnb has a reverse ordering of components being better and worse
+
     """
+    assert isinstance(branches, list), 'branches must be a list'
+    assert isinstance(comp_var_idx, (list, np.ndarray)), 'comp_var_idx must be a list-like'
+    assert isinstance(varis, dict), 'varis must be a dict'
+    assert isinstance(flag_comp_state_order, bool), 'flag_comp_state_order should be either 0 or 1'
+    assert set(comp_var_idx).difference(varis.keys()) == set(), 'varis should contain index of comp_var_idx: {comp_var_idx}'
+
     complete_brs = [x for x in branches if x.is_complete]
-    #nRow = length( complete_brs )
+
+    #FIXME: no_comp = len(comp_var_idx) instead?
     no_comp = len(complete_brs[0].down)
 
     C = np.zeros((len(complete_brs), no_comp + 1))
@@ -84,20 +94,100 @@ def get_cmat(branches, comp_var_idx, varis, flag_comp_state_order=True):
 
             if up_state != down_state:
                 b1 = np.zeros((1, b.shape[1]))
-                b1[down_state-1:up_state-1] = 1
+                b1[down_state - 1:up_state - 1] = 1
 
-                loc = ismember(b1, b)[1]
+                _, loc = ismember(b1, b)
 
                 if any(loc):
                     # conversion to python index
-                    c[j+1] = loc[0] + 1
+                    c[j + 1] = loc[0] + 1
                 else:
+                    #print(f'B of varis[{comp_var_idx[j]}] is updated')
                     b = np.vstack((b, b1))
                     varis[comp_var_idx[j]].B = b
-                    c[j+1] = b.shape[1]
+                    c[j + 1] = b.shape[1]
             else:
-                c[j+1] = up_state
+                c[j + 1] = up_state
 
-        C[irow,:] = c
+        C[irow, :] = c
 
     return C, varis
+
+
+def get_idx(x, flag=False):
+
+    assert isinstance(x, list), 'should be a list'
+    assert all([isinstance(y, Branch) for y in x]), 'should contain an instance of Branch'
+    return [i for i, y in enumerate(x) if y.is_complete is flag]
+
+
+def run_bnb(sys_fn, next_comp_fn, next_state_fn, info, comp_max_states):
+    """
+    return branch
+    Parameters
+    ----------
+    sys_fn
+    next_comp_fn
+    next_state_fn
+    info
+    comp_max_states: list-like
+    """
+    assert callable(sys_fn), 'sys_fn should be a function'
+    assert callable(next_comp_fn), 'next_comp_fn should be a function'
+    assert callable(next_state_fn), 'next_state_fn should be a function'
+    assert isinstance(info, dict), 'info should be a dict'
+    assert isinstance(comp_max_states, list), 'comp_max_states should be a list'
+
+    init_up = comp_max_states
+    init_down = np.ones_like(comp_max_states).tolist() # Assume that the lowest state is 1
+
+    branches = [Branch(down=init_down,
+                       up=init_up,
+                       is_complete=False)]
+
+    incmp_br_idx = get_idx(branches, False)
+
+    while incmp_br_idx:
+        branch_id = incmp_br_idx[0]
+        _branch = branches[branch_id]
+        down = _branch.down
+        up = _branch.up
+
+        down_state, down_val, down_res = sys_fn(down, info)
+        up_state, up_val, up_res = sys_fn(up, info)
+
+        if down_state == up_state:
+            _branch.is_complete = True
+            _branch.down_state = down_state
+            _branch.up_state = up_state
+            _branch.down_val = down_val
+            _branch.up_val = up_val
+            del incmp_br_idx[0]
+
+        else:
+            # matlab vs python index or not
+            cand_next_comp = [info['arcs'][i] for i, (x, y) in enumerate(zip(up, down)) if x > y]
+
+            next_comp = next_comp_fn(cand_next_comp, down_res, up_res, info)
+
+            next_state = next_state_fn(next_comp,
+                                       [_branch.down[next_comp], _branch.up[next_comp]],
+                                       down_res,
+                                       up_res,
+                                       info)
+            branch_down = copy.deepcopy(_branch)
+            branch_down.up[next_comp-1] = next_state
+
+            branch_up = copy.deepcopy(_branch)
+            branch_up.down[next_comp-1] = next_state + 1
+
+            del branches[branch_id]
+
+            branches.append(branch_down)
+            branches.append(branch_up)
+
+            incmp_br_idx = get_idx(branches, False)
+
+    return branches
+
+
