--- run_bnb.py
@@ -1,80 +1,0 @@
-import numpy as np
-import copy
-
-from BNS_JT.branch import Branch
-
-
-def get_idx(x, flag=False):
-
-    assert isinstance(x, list), 'should be a list'
-    assert all([isinstance(y, Branch) for y in x]), 'should contain an instance of Branch'
-    return [i for i, y in enumerate(x) if y.is_complete is flag]
-
-
-def run_bnb(sys_fn, next_comp_fn, next_state_fn, info, comp_max_states):
-    """
-    return branch
-    Parameters
-    ----------
-    sys_fn
-    next_comp_fn
-    next_state_fn
-    info
-    comp_max_states: list-like
-    """
-
-    #ncomp = len(comp_max_states)
-
-    init_up = comp_max_states
-    init_down = np.ones_like(comp_max_states).tolist() # Assume that the lowest state is 1
-
-    branches = [Branch(down=init_down,
-                       up=init_up,
-                       is_complete=False)]
-
-    incmp_br_idx = get_idx(branches, False)
-
-    while incmp_br_idx:
-        branch_id = incmp_br_idx[0]
-        _branch = branches[branch_id]
-        down = _branch.down
-        up = _branch.up
-
-        down_state, down_val, down_res = sys_fn(down, info)
-        up_state, up_val, up_res = sys_fn(up, info)
-
-        if down_state == up_state:
-            _branch.is_complete = True
-            _branch.down_state = down_state
-            _branch.up_state = up_state
-            _branch.down_val = down_val
-            _branch.up_val = up_val
-            del incmp_br_idx[0]
-
-        else:
-            # matlab vs python index or not
-            cand_next_comp = [info['arcs'][i] for i, (x, y) in enumerate(zip(up, down)) if x > y]
-
-            next_comp = next_comp_fn(cand_next_comp, down_res, up_res, info)
-
-            next_state = next_state_fn(next_comp,
-                                       [_branch.down[next_comp], _branch.up[next_comp]],
-                                       down_res,
-                                       up_res,
-                                       info)
-            branch_down = copy.deepcopy(_branch)
-            branch_down.up[next_comp-1] = next_state
-
-            branch_up = copy.deepcopy(_branch)
-            branch_up.down[next_comp-1] = next_state + 1
-
-            del branches[branch_id]
-
-            branches.append(branch_down)
-            branches.append(branch_up)
-
-            incmp_br_idx = get_idx(branches, False)
-
-    return branches
-
-
